chore(code/test): Speed compilation of integration tests by using a single test executable per crate
See https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html for background information
