--- code/crates/core-driver/test-utils/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name        = "informalsystems-malachitebft-core-driver-test-utils"
-description = "Test utilities for the `malachitebft-core-driver`"
-publish     = false
-
-version.workspace      = true
-edition.workspace      = true
-repository.workspace   = true
-license.workspace      = true
-rust-version.workspace = true
-
-
-[dependencies]
-malachitebft-core-driver = { workspace = true }
-malachitebft-core-state-machine = { workspace = true }
-malachitebft-core-types = { workspace = true }
-malachitebft-core-votekeeper = { workspace = true }
-malachitebft-test = { workspace = true }
