+++ code/crates/core-driver/tests/it/extra.rs
@@ -1,11 +1,12 @@
 use malachitebft_core_state_machine::state::State;
 use malachitebft_core_types::{Round, Validity};
 
-use malachitebft_core_driver_test_utils::*;
 use malachitebft_test::utils::validators::make_validators;
 use malachitebft_test::{Height, Proposal, TestContext, ValidatorSet, Value};
 
 use informalsystems_malachitebft_core_driver::{Driver, Input, Output};
+
+use crate::utils::*;
 
 // The following tests are performed:
 // - L49 with commits from current rounds, no locked value, no valid value:
