+++ code/crates/core-driver/tests/it/main.rs
@@ -1,0 +1,4 @@
+pub mod basic;
+pub mod extra;
+
+mod utils;
