+++ code/crates/core-driver/tests/it/utils.rs
@@ -1,9 +1,10 @@
 #![allow(clippy::needless_update)]
 
-use malachitebft_core_driver::{Input, Output};
 use malachitebft_core_state_machine::state::{RoundValue, State, Step};
 use malachitebft_core_types::{NilOrVal, Round, SignedProposal, SignedVote, Timeout, Validity};
 use malachitebft_test::{Address, Height, Proposal, Signature, TestContext, Value, Vote};
+
+use informalsystems_malachitebft_core_driver::{Input, Output};
 
 pub fn new_round_input(round: Round, proposer: Address) -> Input<TestContext> {
     Input::NewRound(Height::new(1), round, proposer)
