+++ code/crates/starknet/test/src/tests.rs
@@ -1,0 +1,7 @@
+pub mod n3f0;
+pub mod n3f0_consensus_mode;
+pub mod n3f0_pubsub_protocol;
+pub mod n3f1;
+pub mod value_sync;
+pub mod vote_sync;
+pub mod wal;
