+++ code/crates/starknet/test/src/tests/n3f0_pubsub_protocol.rs
@@ -1,8 +1,9 @@
 use std::time::Duration;
 
 use bytesize::ByteSize;
-use informalsystems_malachitebft_starknet_test::{init_logging, TestBuilder, TestParams};
 use malachitebft_config::{GossipSubConfig, PubSubProtocol};
+
+use crate::{init_logging, TestBuilder, TestParams};
 
 async fn run_test(params: TestParams) {
     init_logging(module_path!());
