+++ code/crates/starknet/test/src/tests/vote_sync.rs
@@ -1,7 +1,11 @@
 use std::time::Duration;
 
-use informalsystems_malachitebft_starknet_test::{init_logging, TestBuilder, TestParams};
 use malachitebft_config::ValuePayload;
+
+use crate::{init_logging, TestBuilder, TestParams};
+
+// NOTE: These tests are very similar to the Sync tests, with the difference that
+//       all nodes have the same voting power and therefore get stuck when one of them dies.
 
 pub async fn crash_restart_from_start(params: TestParams) {
     init_logging(module_path!());
@@ -10,34 +14,21 @@ pub async fn crash_restart_from_start(params: TestParams) {
 
     let mut test = TestBuilder::<()>::new();
 
-    // Node 1 starts with 10 voting power.
-    test.add_node()
-        .with_voting_power(10)
-        .start()
-        // Wait until it reaches height 10
-        .wait_until(HEIGHT)
-        // Record a successful test for this node
-        .success();
+    test.add_node().start().wait_until(HEIGHT).success();
+    test.add_node().start().wait_until(HEIGHT).success();
 
-    // Node 2 starts with 10 voting power, in parallel with node 1 and with the same behaviour
     test.add_node()
-        .with_voting_power(10)
         .start()
-        .wait_until(HEIGHT)
-        .success();
-
-    // Node 3 starts with 5 voting power, in parallel with node 1 and 2.
-    test.add_node()
-        .with_voting_power(5)
-        .start()
-        // Wait until the node reaches height 2...
-        .wait_until(2)
-        // ...and then kills it
+        // Wait until the node reaches height 4...
+        .wait_until(4)
+        // ...then kill it
         .crash()
         // Reset the database so that the node has to do Sync from height 1
         .reset_db()
         // After that, it waits 5 seconds before restarting the node
         .restart_after(Duration::from_secs(5))
+        // Expect a vote set request for height 4
+        .expect_vote_set_request(4)
         // Wait until the node reached the expected height
         .wait_until(HEIGHT)
         // Record a successful test for this node
@@ -48,6 +39,7 @@ pub async fn crash_restart_from_start(params: TestParams) {
             Duration::from_secs(60), // Timeout for the whole test
             TestParams {
                 enable_sync: true, // Enable Sync
+                timeout_step: Duration::from_secs(5),
                 ..params
             },
         )
@@ -92,24 +84,16 @@ pub async fn crash_restart_from_latest() {
 
     let mut test = TestBuilder::<()>::new();
 
+    test.add_node().start().wait_until(HEIGHT).success();
+    test.add_node().start().wait_until(HEIGHT).success();
     test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT)
-        .success();
-    test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT)
-        .success();
-
-    test.add_node()
-        .with_voting_power(5)
         .start()
         .wait_until(2)
         .crash()
         // We do not reset the database so that the node can restart from the latest height
         .restart_after(Duration::from_secs(5))
+        // Expect a vote set request for height 2
+        .expect_vote_set_request(2)
         .wait_until(HEIGHT)
         .success();
 
@@ -118,48 +102,7 @@ pub async fn crash_restart_from_latest() {
             Duration::from_secs(60),
             TestParams {
                 enable_sync: true,
-                ..Default::default()
-            },
-        )
-        .await
-}
-
-#[tokio::test]
-pub async fn aggressive_pruning() {
-    init_logging(module_path!());
-
-    const HEIGHT: u64 = 15;
-
-    let mut test = TestBuilder::<()>::new();
-
-    // Node 1 starts with 10 voting power.
-    test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT)
-        .success();
-    test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT)
-        .success();
-
-    test.add_node()
-        .with_voting_power(5)
-        .start()
-        .wait_until(2)
-        .crash()
-        .reset_db()
-        .restart_after(Duration::from_secs(5))
-        .wait_until(HEIGHT)
-        .success();
-
-    test.build()
-        .run_with_custom_config(
-            Duration::from_secs(60), // Timeout for the whole test
-            TestParams {
-                enable_sync: true,     // Enable Sync
-                max_retain_blocks: 10, // Prune blocks older than 10
+                timeout_step: Duration::from_secs(5),
                 ..Default::default()
             },
         )
@@ -168,33 +111,26 @@ pub async fn aggressive_pruning() {
 
 #[tokio::test]
 pub async fn start_late() {
-    const HEIGHT: u64 = 5;
+    init_logging(module_path!());
 
+    const HEIGHT: u64 = 5;
     let mut test = TestBuilder::<()>::new();
 
+    test.add_node().start().wait_until(HEIGHT * 2).success();
+    test.add_node().start().wait_until(HEIGHT * 2).success();
     test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT * 2)
-        .success();
-
-    test.add_node()
-        .with_voting_power(10)
-        .start()
-        .wait_until(HEIGHT * 2)
-        .success();
-
-    test.add_node()
-        .with_voting_power(5)
         .start_after(1, Duration::from_secs(10))
+        // Expect a vote set request for height 1
+        .expect_vote_set_request(1)
         .wait_until(HEIGHT)
         .success();
 
     test.build()
         .run_with_custom_config(
-            Duration::from_secs(30),
+            Duration::from_secs(60),
             TestParams {
                 enable_sync: true,
+                timeout_step: Duration::from_secs(5),
                 ..Default::default()
             },
         )
