+++ code/crates/starknet/test/src/tests/wal.rs
@@ -9,9 +9,7 @@ use malachitebft_core_types::SignedVote;
 use malachitebft_engine::util::events::Event;
 use malachitebft_starknet_host::types::MockContext;
 
-use informalsystems_malachitebft_starknet_test::{
-    init_logging, HandlerResult, TestBuilder, TestParams,
-};
+use crate::{init_logging, HandlerResult, TestBuilder, TestParams};
 
 #[tokio::test]
 async fn proposer_crashes_after_proposing_parts_only() {
