+++ code/crates/test/mbt/src/lib.rs
@@ -3,3 +3,6 @@ pub mod deserializers;
 pub mod types;
 pub mod utils;
 pub mod votekeeper;
+
+#[cfg(test)]
+pub mod tests;
