+++ code/crates/wal/src/lib.rs
@@ -2,7 +2,6 @@
 
 //! Write-Ahead Log (WAL) implementation
 
-mod ext;
 mod file;
 mod storage;
 mod version;
@@ -12,3 +11,7 @@ pub mod log;
 pub use file::{Log, LogEntry, LogIter};
 pub use storage::Storage;
 pub use version::Version;
+
+// For use in tests
+#[doc(hidden)]
+pub mod ext;
