+++ code/crates/wal/tests/it/main.rs
@@ -1,0 +1,7 @@
+pub mod basic;
+pub mod corruption;
+pub mod crashes;
+pub mod stress;
+
+#[cfg(all(feature = "compression", not(feature = "force-compression")))]
+pub mod compression;
